♻️ 削除機能のリファクタリング
